Move generic cli logic to substrate-relay-helper
