--- relay-primitives/cli/Cargo.toml
@@ -1,57 +1,0 @@
-[package]
-name = "brp-cli"
-version = "1.2.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow = "1.0"
-async-std = "1.9.0"
-async-trait = "0.1"
-codec = { package = "parity-scale-codec", version = "3.1.5" }
-env_logger = "0.11"
-futures = "0.3.30"
-hex = "0.4"
-log = { workspace = true }
-num-format = "0.4"
-num-traits = "0.2"
-rbtag = "0.3"
-structopt = "0.3"
-signal-hook = "0.3.15"
-signal-hook-async-std = "0.2.2"
-strum = { version = "0.26.2", features = ["derive"] }
-
-# Bridge dependencies
-bp-bridge-hub-polkadot = { path = "../../primitives/chain-bridge-hub-polkadot" }
-bp-bridge-hub-rococo = { path = "../../primitives/chain-bridge-hub-rococo" }
-bp-header-chain = { path = "../../primitives/header-chain" }
-bp-messages = { path = "../../primitives/messages" }
-bp-parachains = { path = "../../primitives/parachains" }
-bp-polkadot-bulletin = { path = "../../primitives/chain-polkadot-bulletin" }
-bp-polkadot = { path = "../../primitives/chain-polkadot" }
-bp-polkadot-core = { path = "../../primitives/polkadot-core" }
-bp-rococo = { path = "../../primitives/chain-rococo" }
-bp-runtime = { path = "../../primitives/runtime" }
-bridge-runtime-common = { path = "../../bin/runtime-common" }
-pallet-bridge-parachains = { path = "../../modules/parachains" }
-parachains-relay = { path = "../../relays/parachains" }
-relay-substrate-client = { path = "../../relays/client-substrate" }
-relay-utils = { path = "../../relays/utils" }
-substrate-relay-helper = { path = "../../relays/lib-substrate-relay" }
-
-# Substrate Dependencies
-
-frame-support = { git = "https://github.com/paritytech/polkadot-sdk", branch = "master" }
-sp-core = { git = "https://github.com/paritytech/polkadot-sdk", branch = "master" }
-sp-runtime = { git = "https://github.com/paritytech/polkadot-sdk", branch = "master" }
-
-[dev-dependencies]
-bp-test-utils = { path = "../../primitives/test-utils" }
-hex-literal = "0.4"
-sp-keyring = { git = "https://github.com/paritytech/polkadot-sdk", branch = "master" }
-tempfile = "3.10"
-finality-grandpa = { version = "0.16.2" }
