+++ relays/bin-substrate/src/bridges/kusama_polkadot/bridge_hub_polkadot_messages_to_bridge_hub_kusama.rs
@@ -16,10 +16,13 @@
 
 //! BridgeHubPolkadot-to-BridgeHubKusama messages sync entrypoint.
 
-use brp_cli::bridge::{CliBridgeBase, MessagesCliBridge};
 use relay_bridge_hub_kusama_client::BridgeHubKusama;
 use relay_bridge_hub_polkadot_client::BridgeHubPolkadot;
-use substrate_relay_helper::{messages_lane::SubstrateMessageLane, UtilityPalletBatchCallBuilder};
+use substrate_relay_helper::{
+	cli::bridge::{CliBridgeBase, MessagesCliBridge},
+	messages_lane::SubstrateMessageLane,
+	UtilityPalletBatchCallBuilder,
+};
 
 /// BridgeHubPolkadot-to-BridgeHubKusama messages bridge.
 pub struct BridgeHubPolkadotToBridgeHubKusamaMessagesCliBridge {}
