+++ relays/bin-substrate/src/bridges/kusama_polkadot/kusama_parachains_to_bridge_hub_polkadot.rs
@@ -17,10 +17,10 @@
 //! Kusama-to-BridgeHubPolkadot parachains sync entrypoint.
 
 use bp_polkadot_core::parachains::{ParaHash, ParaHeadsProof, ParaId};
-use brp_cli::bridge::{CliBridgeBase, MessagesCliBridge, ParachainToRelayHeadersCliBridge};
 use relay_substrate_client::{CallOf, HeaderIdOf};
-use substrate_relay_helper::parachains::{
-	SubmitParachainHeadsCallBuilder, SubstrateParachainsPipeline,
+use substrate_relay_helper::{
+	cli::bridge::{CliBridgeBase, MessagesCliBridge, ParachainToRelayHeadersCliBridge},
+	parachains::{SubmitParachainHeadsCallBuilder, SubstrateParachainsPipeline},
 };
 
 /// Kusama-to-BridgeHubPolkadot parachain sync description.
