+++ relays/bin-substrate/src/bridges/polkadot_bulletin/bridge_hub_polkadot_messages_to_polkadot_bulletin.rs
@@ -16,10 +16,13 @@
 
 //! BridgeHubPolkadot-to-PolkadotBulletin messages sync entrypoint.
 
-use brp_cli::bridge::{CliBridgeBase, MessagesCliBridge};
 use relay_bridge_hub_polkadot_client::BridgeHubPolkadot;
 use relay_polkadot_bulletin_client::PolkadotBulletin;
-use substrate_relay_helper::{messages_lane::SubstrateMessageLane, UtilityPalletBatchCallBuilder};
+use substrate_relay_helper::{
+	cli::bridge::{CliBridgeBase, MessagesCliBridge},
+	messages_lane::SubstrateMessageLane,
+	UtilityPalletBatchCallBuilder,
+};
 
 /// BridgeHubPolkadot-to-PolkadotBulletin messages bridge.
 pub struct BridgeHubPolkadotToPolkadotBulletinMessagesCliBridge {}
