+++ relays/bin-substrate/src/bridges/polkadot_bulletin/polkadot_bulletin_headers_to_bridge_hub_polkadot.rs
@@ -23,7 +23,7 @@ use substrate_relay_helper::{
 	finality_base::{engine::Grandpa as GrandpaFinalityEngine, SubstrateFinalityPipeline},
 };
 
-use brp_cli::bridge::{
+use substrate_relay_helper::cli::bridge::{
 	CliBridgeBase, MessagesCliBridge, RelayToRelayEquivocationDetectionCliBridge,
 	RelayToRelayHeadersCliBridge,
 };
