+++ relays/bin-substrate/src/bridges/polkadot_bulletin/polkadot_bulletin_messages_to_bridge_hub_polkadot.rs
@@ -16,10 +16,13 @@
 
 //! PolkadotBulletin-to-BridgeHubPolkadot messages sync entrypoint.
 
-use brp_cli::bridge::{CliBridgeBase, MessagesCliBridge};
 use relay_bridge_hub_polkadot_client::BridgeHubPolkadot;
 use relay_polkadot_bulletin_client::PolkadotBulletin;
-use substrate_relay_helper::{messages_lane::SubstrateMessageLane, UtilityPalletBatchCallBuilder};
+use substrate_relay_helper::{
+	cli::bridge::{CliBridgeBase, MessagesCliBridge},
+	messages_lane::SubstrateMessageLane,
+	UtilityPalletBatchCallBuilder,
+};
 
 /// PolkadotBulletin-to-BridgeHubPolkadot messages bridge.
 pub struct PolkadotBulletinToBridgeHubPolkadotMessagesCliBridge {}
