+++ relays/bin-substrate/src/bridges/rococo_bulletin/mod.rs
@@ -23,7 +23,8 @@ use bp_runtime::{
 };
 use frame_support::pallet_prelude::Weight;
 use relay_substrate_client::{
-	ChainWithRuntimeVersion, Error as SubstrateError, SignParam, SimpleRuntimeVersion, UnsignedTransaction,
+	ChainWithRuntimeVersion, Error as SubstrateError, SignParam, SimpleRuntimeVersion,
+	UnsignedTransaction,
 };
 use sp_core::storage::StorageKey;
 use std::time::Duration;
