+++ relays/bin-substrate/src/bridges/rococo_bulletin/rococo_bulletin_headers_to_bridge_hub_rococo.rs
@@ -25,7 +25,7 @@ use substrate_relay_helper::{
 	finality_base::{engine::Grandpa as GrandpaFinalityEngine, SubstrateFinalityPipeline},
 };
 
-use brp_cli::bridge::{
+use substrate_relay_helper::cli::bridge::{
 	CliBridgeBase, MessagesCliBridge, RelayToRelayEquivocationDetectionCliBridge,
 	RelayToRelayHeadersCliBridge,
 };
