+++ relays/bin-substrate/src/bridges/rococo_bulletin/rococo_bulletin_messages_to_bridge_hub_rococo.rs
@@ -17,9 +17,12 @@
 //! RococoBulletin-to-BridgeHubRococo messages sync entrypoint.
 
 use super::BridgeHubRococoAsBridgeHubPolkadot;
-use brp_cli::bridge::{CliBridgeBase, MessagesCliBridge};
 use relay_polkadot_bulletin_client::PolkadotBulletin as RococoBulletin;
-use substrate_relay_helper::{messages_lane::SubstrateMessageLane, UtilityPalletBatchCallBuilder};
+use substrate_relay_helper::{
+	cli::bridge::{CliBridgeBase, MessagesCliBridge},
+	messages_lane::SubstrateMessageLane,
+	UtilityPalletBatchCallBuilder,
+};
 
 /// RococoBulletin-to-BridgeHubRococo messages bridge.
 pub struct RococoBulletinToBridgeHubRococoMessagesCliBridge {}
