+++ relays/bin-substrate/src/bridges/rococo_bulletin/rococo_parachains_to_rococo_bulletin.rs
@@ -20,9 +20,9 @@ use super::{BridgeHubRococoAsBridgeHubPolkadot, RococoAsPolkadot};
 
 use bp_polkadot_core::parachains::{ParaHash, ParaHeadsProof, ParaId};
 use bp_runtime::Chain;
-use brp_cli::bridge::{CliBridgeBase, MessagesCliBridge, ParachainToRelayHeadersCliBridge};
 use relay_substrate_client::{CallOf, HeaderIdOf};
 use substrate_relay_helper::{
+	cli::bridge::{CliBridgeBase, MessagesCliBridge, ParachainToRelayHeadersCliBridge},
 	messages_lane::MessagesRelayLimits,
 	parachains::{SubmitParachainHeadsCallBuilder, SubstrateParachainsPipeline},
 };
