+++ relays/bin-substrate/src/bridges/rococo_westend/rococo_parachains_to_bridge_hub_westend.rs
@@ -17,10 +17,10 @@
 //! Westend-to-Rococo parachains sync entrypoint.
 
 use bp_polkadot_core::parachains::{ParaHash, ParaHeadsProof, ParaId};
-use brp_cli::bridge::{CliBridgeBase, MessagesCliBridge, ParachainToRelayHeadersCliBridge};
 use relay_substrate_client::{CallOf, HeaderIdOf};
-use substrate_relay_helper::parachains::{
-	SubmitParachainHeadsCallBuilder, SubstrateParachainsPipeline,
+use substrate_relay_helper::{
+	cli::bridge::{CliBridgeBase, MessagesCliBridge, ParachainToRelayHeadersCliBridge},
+	parachains::{SubmitParachainHeadsCallBuilder, SubstrateParachainsPipeline},
 };
 
 /// BridgeHub-to-BridgeHub parachain sync description.
