+++ relays/bin-substrate/src/bridges/rococo_westend/westend_headers_to_bridge_hub_rococo.rs
@@ -23,7 +23,7 @@ use substrate_relay_helper::{
 	finality_base::{engine::Grandpa as GrandpaFinalityEngine, SubstrateFinalityPipeline},
 };
 
-use brp_cli::bridge::{
+use substrate_relay_helper::cli::bridge::{
 	CliBridgeBase, RelayToRelayEquivocationDetectionCliBridge, RelayToRelayHeadersCliBridge,
 };
