+++ relays/client-bridge-hub-westend/src/lib.rs
@@ -23,9 +23,9 @@ use bp_polkadot_core::SuffixedCommonTransactionExtensionExt;
 use codec::Encode;
 use relay_substrate_client::{
 	calls::UtilityCall as MockUtilityCall, Chain, ChainWithBalances, ChainWithMessages,
-	ChainWithTransactions, ChainWithUtilityPallet, ChainWithRuntimeVersion, Error as SubstrateError,
-	MockedRuntimeUtilityPallet, SignParam, SimpleRuntimeVersion, UnderlyingChainProvider,
-	UnsignedTransaction,
+	ChainWithRuntimeVersion, ChainWithTransactions, ChainWithUtilityPallet,
+	Error as SubstrateError, MockedRuntimeUtilityPallet, SignParam, SimpleRuntimeVersion,
+	UnderlyingChainProvider, UnsignedTransaction,
 };
 use sp_core::{storage::StorageKey, Pair};
 use sp_runtime::{generic::SignedPayload, traits::IdentifyAccount};
