+++ relays/client-substrate/src/lib.rs
@@ -35,9 +35,9 @@ use std::time::Duration;
 pub use crate::{
 	chain::{
 		AccountKeyPairOf, BlockWithJustification, CallOf, Chain, ChainWithBalances,
-		ChainWithGrandpa, ChainWithMessages, ChainWithTransactions, ChainWithUtilityPallet,
-		ChainWithRuntimeVersion, FullRuntimeUtilityPallet, MockedRuntimeUtilityPallet, Parachain, RelayChain,
-		SignParam, TransactionStatusOf, UnsignedTransaction, UtilityPallet,
+		ChainWithGrandpa, ChainWithMessages, ChainWithRuntimeVersion, ChainWithTransactions,
+		ChainWithUtilityPallet, FullRuntimeUtilityPallet, MockedRuntimeUtilityPallet, Parachain,
+		RelayChain, SignParam, TransactionStatusOf, UnsignedTransaction, UtilityPallet,
 	},
 	client::{
 		is_ancient_block, ChainRuntimeVersion, Client, OpaqueGrandpaAuthoritiesSet,
