+++ relays/lib-substrate-relay/Cargo.toml
@@ -10,14 +10,17 @@ workspace = true
 
 [dependencies]
 anyhow = "1.0"
-thiserror = { workspace = true }
 async-std = "1.9.0"
 async-trait = "0.1"
 codec = { package = "parity-scale-codec", version = "3.1.5" }
 futures = "0.3.30"
 hex = "0.4"
-num-traits = "0.2"
 log = { workspace = true }
+num-traits = "0.2"
+rbtag = "0.3"
+structopt = "0.3"
+strum = { version = "0.26.2", features = ["derive"] }
+thiserror = { workspace = true }
 
 # Bridge dependencies
